refactor: Organize down tests and move task to its own file.
